Fix bug and use clearer expanded columns
The columns now reflect the location of the value in the JSON. I hate to
change the column names from under those using the SDK, but we've
received feedback that it's hard to tell how to read these column names.

The minor fix is when a valueCodeableConcept does not have coding
values.